+++ Lib.Test/ChromePathFinderTest.cs
@@ -60,36 +60,35 @@ namespace Lib.Test
         [Fact]
         void ReturnsWindowsPathIfNotUnixFs()
         {
-            var chromePath = BrowserPathFinder.GetBrowserPath(new FakeFs(false, @"C:/Program Files (x86)/Google/Chrome/Application/chrome.exe"));
+            var chromePath = BrowserPathFinder.GetBrowserPath(new FakeFs(false, @"C:/Program Files (x86)/Google/Chrome/Application/chrome.exe"), true);
             Assert.Equal(@"C:/Program Files (x86)/Google/Chrome/Application/chrome.exe", chromePath);
         }
 
         [Fact]
         void ReturnsLinuxPathIfIsUnixFs()
         {
-            var chromePath = BrowserPathFinder.GetBrowserPath(new FakeFs(true, "/opt/google/chrome/google-chrome"));
+            var chromePath = BrowserPathFinder.GetBrowserPath(new FakeFs(true, "/opt/google/chrome/google-chrome"), true);
             Assert.Equal("/opt/google/chrome/google-chrome", chromePath);
         }
 
         [Fact]
         void ReturnsLinuxChromiumPathIfIsUnixFsAndChromeNotInstalled()
         {
-            var chromePath = BrowserPathFinder.GetBrowserPath(new FakeFs(true, "/usr/bin/chromium"));
+            var chromePath = BrowserPathFinder.GetBrowserPath(new FakeFs(true, "/usr/bin/chromium"), true);
             Assert.Equal("/usr/bin/chromium", chromePath);
         }
 
         [Fact]
         void ReturnsLinuxChromiumBrowserPathIfIsUnixFsAndChromeNotInstalled()
         {
-            var chromePath = BrowserPathFinder.GetBrowserPath(new FakeFs(true, "/usr/bin/chromium-browser"));
+            var chromePath = BrowserPathFinder.GetBrowserPath(new FakeFs(true, "/usr/bin/chromium-browser"), true);
             Assert.Equal("/usr/bin/chromium-browser", chromePath);
         }
 
         [Fact]
-        void ThrowsExceptionWhenChromeNorChromiumIsNotFound()
+        void ReturnsNullWhenChromeNorChromiumIsNotFound()
         {
-            Exception ex = Assert.Throws<Exception>(() => BrowserPathFinder.GetBrowserPath(new FakeFs(true)));
-            Assert.Contains("Browser not found", ex.Message);
+            Assert.Null(BrowserPathFinder.GetBrowserPath(new FakeFs(true), true));
         }
     }